Move aiTimeToSampleSelector outside of AbcAPI
