+++ Source/abci/Importer/AlembicImporter.cpp
@@ -19,16 +19,6 @@ abciAPI void aiTimeToSampleSelector(double time, aiTimeToSampleSelectorInfo* dst
     dst->requestedTimeIndexType = ss.getRequestedTimeIndexType();
 }
 
-abcSampleSelector aiTimeToSampleSelector(double time)
-{
-    return abcSampleSelector(time, abcSampleSelector::kFloorIndex);
-}
-
-abcSampleSelector aiIndexToSampleSelector(int64_t index)
-{
-    return abcSampleSelector(index);
-}
-
 abciAPI void aiCleanup()
 {
 }
