+++ Source/abci/Importer/aiSampleSelector.cpp
@@ -1,0 +1,11 @@
+#include "aiSampleSelector.h"
+
+abcSampleSelector aiTimeToSampleSelector(double time)
+{
+    return abcSampleSelector(time, abcSampleSelector::kFloorIndex);
+}
+
+abcSampleSelector aiIndexToSampleSelector(int64_t index)
+{
+    return abcSampleSelector(index);
+}
